+++ modules/hanna-react/CHANGELOG.md
@@ -5,6 +5,7 @@
 - ... <!-- Add new lines here. -->
 - `MainÞMenu2`:
   - feat: Pass `openMenu()` prop to `MainMenu2CustomItem` callbacks
+  - feat: Allow rich-text (JSX) `label`s for menu items
   - fix: links with empty-string `href` not rendered server-side
 
 ## 0.10.133